pyproject.toml
[tool.poetry]
name = "pygym"
version = "0.1.0"
description = "Python Gym - Python coding challenges"
authors = ["Andrew Kurauchi <andrew.kurauchi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.3"
django-extensions = "^3.1.0"
requests = "^2.25.0"
Markdown = "^3.3.3"
djangorestframework = "^3.12.2"
strtest = "^0.0.4"
django-cors-headers = "^3.7.0"
python-dotenv = "^0.15.0"
gunicorn = "^20.0.4"
django-storages = "^1.11.1"
boto3 = "^1.17.6"

[tool.poetry.dev-dependencies]
coverage = "^5.3"
pytest = "^6.1.2"
uvicorn = "^0.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
